% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_transformer.R
\name{load_dependencies}
\alias{load_dependencies}
\title{Load dependencies for an individual model}
\usage{
load_dependencies(model_path, language_pair = "fr-de", transformers)
}
\arguments{
\item{model_path}{Path to the model. If null, the remote is used.}

\item{transformers}{}
}
\description{
Load dependencies for an individual model
}
