% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_transformer.R
\name{setup_translator}
\alias{setup_translator}
\title{Setup_translator}
\usage{
setup_translator(language_pairs = "fr-de", model_path = NULL)
}
\arguments{
\item{language_out}{}
}
\description{
Setup_translator
}
