% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_translate.R
\name{translate}
\alias{translate}
\title{Translate}
\usage{
translate(input_text, dependencies, language_pair, verbose = TRUE)
}
\arguments{
\item{input_text}{String of the text to be translated}

\item{dependencies}{Object storing the model and tokenizer for `language_pair.`.
Output of [setup_translator()]}

\item{language_pair}{String, specifying the pair of languages to be translated.
E.g. `"de-fr"`}

\item{verbose}{Logical, should messages be printed to the console?}
}
\value{
The translation of input_text
}
\description{
This function translates a particular string.
}
\examples{
\dontrun{translate(input_text = "Haus", dependencies = deps, language_pair = "de-fr")}
}
