% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_translator.R
\name{load_dependencies}
\alias{load_dependencies}
\title{Load dependencies for an individual model}
\usage{
load_dependencies(language_pair = "fr-de", transformers, model_path = NULL)
}
\arguments{
\item{language_pair}{String, specifying the pair of languages to be
translated. E.g. `"de-fr"`}

\item{transformers}{The transformers module}

\item{model_path}{Path to the model. If null, the default cache folder
(e.g. `~/.cache/huggingface/hub/`) is used.}
}
\value{
A list containing the model and tokenizer for one language pair
}
\description{
This function loads the tokenizer and model for a specific language pair.
}
\examples{
\dontrun{load_dependencies(language_pair = "fr-de", transformers = .python_modules$transformers)}
}
