% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_transformer.R
\name{setup_translator}
\alias{setup_translator}
\title{Setup_translator}
\usage{
setup_translator(language_pairs = "de-fr", model_path = NULL)
}
\arguments{
\item{language_pairs}{Character vector, specifying the pairs of languages to be
translated. E.g. `"de-fr"` or `c("de-fr", "en-fr")`}

\item{model_path}{Path to the model. If null, the default cache folder
(e.g. `~/.cache/huggingface/hub/`) is used.}
}
\value{
A list containing the model and tokenizer for the language pairs
}
\description{
Setup_translator
}
\examples{
\dontrun{setup_translator()}
}
