% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_transformer.R
\name{translate}
\alias{translate}
\title{Translate}
\usage{
translate(
  dependencies,
  language_pair,
  input_text,
  character_cut = 1200,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Translate
}
